// packages/electron/src/models/putMetricDataRequest.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface PutMetricDataRequest {
  requestMetadata?: {
    service?: string;
    created?: string;
    namespace?: string;
    description?: string;
    version?: string;
    token?: string;
    [k: string]: unknown;
  };
  metrics?: {
    timeStamp?: string;
    unit?: string;
    metricName?: string;
    value?: number;
    dimensions?: {
      Value?: string;
      Name?: string;
      [k: string]: unknown;
    }[];
    [k: string]: unknown;
  }[];
  [k: string]: unknown;
}